requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
jupyter>=1.0.0
notebook>=6.4.0
fredapi>=0.5.0
statsmodels>=0.13.0
plotly>=5.3.0
requests>=2.26.0
python-dotenv>=0.19.0
joblib>=1.1.0
